add perf test and use static casting

+++ driver/mloMatrixDiagHost.hpp
@@ -32,14 +32,16 @@
 #include <miopen/miopen.h>
 #include <miopen/tensor.hpp>
 
-inline int GetOffset(int max_diag_len, int d, int M, int N, miopenMatrixDiagAlignMode_t align)
+inline int64_t
+GetOffset(int64_t max_diag_len, int64_t d, int64_t M, int64_t N, miopenMatrixDiagAlignMode_t align)
 {
     if(((align == MIOPEN_MATRIX_ALIGN_RIGHT_LEFT || align == MIOPEN_MATRIX_ALIGN_RIGHT_RIGHT) &&
         d >= 0) ||
        ((align == MIOPEN_MATRIX_ALIGN_LEFT_RIGHT || align == MIOPEN_MATRIX_ALIGN_RIGHT_RIGHT) &&
         d <= 0))
     {
-        return max_diag_len - std::min(N - std::max(d, 0), M + std::min(d, 0));
+        return max_diag_len - std::min(N - std::max(d, static_cast<int64_t>(0)),
+                                       M + std::min(d, static_cast<int64_t>(0)));
     }
     return 0;
 }
@@ -63,20 +65,22 @@ int32_t mloMatrixSetDiagForwardRunHost(const miopenTensorDescriptor_t inputDesc,
     auto M        = outShape[outShape.size() - 2];
     auto N        = outShape[outShape.size() - 1];
 
-    int max_diag_len = std::min(M + std::min(k1, 0L), N + std::min(-k0, 0L));
-    T input_val      = (inSize > 0 ? input[0] : static_cast<T>(0));
+    int64_t max_diag_len = std::min(M + std::min(k1, static_cast<int64_t>(0)),
+                                    N + std::min(-k0, static_cast<int64_t>(0)));
+    T input_val          = (inSize > 0 ? input[0] : static_cast<T>(0));
 
     par_ford(outSize)([&](size_t gid) {
-        int batch_id = gid / M / N;
-        int m        = (gid / N) % M;
-        int n        = gid % N;
+        int64_t batch_id = gid / M / N;
+        int64_t m        = (gid / N) % M;
+        int64_t n        = gid % N;
 
         T val;
         if(k0 == k1)
         {
             if(n - m == k1)
             {
-                int diag_id = batch_id * max_diag_len + n - std::max(k1, 0L);
+                int64_t diag_id =
+                    batch_id * max_diag_len + n - std::max(k1, static_cast<int64_t>(0));
                 if(is_fwd)
                     val = diag[diag_id];
                 else
@@ -92,14 +96,14 @@ int32_t mloMatrixSetDiagForwardRunHost(const miopenTensorDescriptor_t inputDesc,
         }
         else
         {
-            int d = n - m;
+            int64_t d = n - m;
             if(k0 <= d && d <= k1)
             {
-                int num_diags     = k1 - k0 + 1;
-                int diag_index    = k1 - d;
-                int offset        = GetOffset(max_diag_len, d, M, N, align);
-                int index_in_diag = n - std::max(d, 0) + offset;
-                int diag_id =
+                int64_t num_diags     = k1 - k0 + 1;
+                int64_t diag_index    = k1 - d;
+                int64_t offset        = GetOffset(max_diag_len, d, M, N, align);
+                int64_t index_in_diag = n - std::max(d, static_cast<int64_t>(0)) + offset;
+                int64_t diag_id =
                     batch_id * num_diags * max_diag_len + diag_index * max_diag_len + index_in_diag;
                 if(is_fwd)
                     val = diag[diag_id];
@@ -138,21 +142,22 @@ int32_t mloMatrixDiagPartForwardRunHost(const miopenTensorDescriptor_t inputDesc
     auto M       = inShape[inShape.size() - 2];
     auto N       = inShape[inShape.size() - 1];
 
-    int max_diag_len = std::min(M + std::min(k1, 0L), N + std::min(-k0, 0L));
-    T padding_val    = padSize > 0 ? pad[0] : static_cast<T>(0);
+    int64_t max_diag_len = std::min(M + std::min(k1, static_cast<int64_t>(0)),
+                                    N + std::min(-k0, static_cast<int64_t>(0)));
+    T padding_val        = padSize > 0 ? pad[0] : static_cast<T>(0);
 
     par_ford(outSize)([&](size_t gid) {
         T val;
         if(k0 == k1)
         {
-            int batch_id = gid / max_diag_len;
-            int n        = gid % max_diag_len;
-            int x        = std::max(k1, 0L);
-            int y        = std::max(-k1, 0L);
+            int64_t batch_id = gid / max_diag_len;
+            int64_t n        = gid % max_diag_len;
+            int64_t x        = std::max(k1, static_cast<int64_t>(0));
+            int64_t y        = std::max(-k1, static_cast<int64_t>(0));
             if(0 <= n + y && n + y < M && 0 <= n + x && n + x < N)
             {
-                int input_id = batch_id * M * N + (n + y) * N + n + x;
-                val          = input[input_id];
+                int64_t input_id = batch_id * M * N + (n + y) * N + n + x;
+                val              = input[input_id];
             }
             else
             {
@@ -161,19 +166,19 @@ int32_t mloMatrixDiagPartForwardRunHost(const miopenTensorDescriptor_t inputDesc
         }
         else
         {
-            int num_diags = k1 - k0 + 1;
-            int batch_id  = gid / num_diags / max_diag_len;
-            int m         = (gid / max_diag_len) % num_diags;
-            int n         = gid % max_diag_len;
-            int d         = k1 - m;
-            int offset    = GetOffset(max_diag_len, d, M, N, align);
-            int y         = std::max(-d, 0) - offset;
-            int x         = std::max(d, 0) - offset;
+            int64_t num_diags = k1 - k0 + 1;
+            int64_t batch_id  = gid / num_diags / max_diag_len;
+            int64_t m         = (gid / max_diag_len) % num_diags;
+            int64_t n         = gid % max_diag_len;
+            int64_t d         = k1 - m;
+            int64_t offset    = GetOffset(max_diag_len, d, M, N, align);
+            int64_t y         = std::max(-d, static_cast<int64_t>(0)) - offset;
+            int64_t x         = std::max(d, static_cast<int64_t>(0)) - offset;
 
             if(0 <= n + y && n + y < M && 0 <= n + x && n + x < N)
             {
-                int input_id = batch_id * M * N + (n + y) * N + n + x;
-                val          = input[input_id];
+                int64_t input_id = batch_id * M * N + (n + y) * N + n + x;
+                val              = input[input_id];
             }
             else
             {
